Fix C/Fortran char length type mismatch: use size_t not f77int
On 64-bit systems, Intel Fortran passes hidden character length
arguments as 8-byte integers (size_t/long), but bufrlib.h declared
them as f77int (int, 32-bit). This caused the actual length to be
read from the wrong memory location, producing large negative values.

Fix: Changed all hidden character length parameters in bufrlib.h
prototypes from f77int to size_t, and updated all C callers to pass
(size_t)1 instead of plain 1.

Files modified:
- bufrlib.h: Changed hidden char length params to size_t in prototypes
  for ichkstr, igetntbi, ipkm, iupm, nemtab, numtbd, rbytes, stntbi
- stseq.c: Pass (size_t)1 for TAB length in numtbd/nemtab/igetntbi calls
- restd.c: Pass (size_t)1 for TAB length in numtbd calls

+++ NCEP_bufr/stseq.c
@@ -88,7 +88,7 @@ void stseq( f77int *lun, f77int *irepct, f77int *idn, char nemo[8],
 **  Is *idn already listed as an entry in the internal Table D?
 **  If so, then there's no need to proceed any further.
 */
-    numtbd( lun, idn, nemo2, tab, &iret, sizeof( nemo2 ), 1 );
+    numtbd( lun, idn, nemo2, tab, &iret, sizeof( nemo2 ), (size_t)1 );
     if ( ( iret > 0 ) && ( tab[0] == 'D' ) ) return;
 
 /*
@@ -96,7 +96,7 @@ void stseq( f77int *lun, f77int *irepct, f77int *idn, char nemo[8],
 */
     tab[0] = 'D';
     tab[1] = '\0';
-    nd = igetntbi( lun, tab, 1 );
+    nd = igetntbi( lun, tab, (size_t)1 );
     cadn30( idn, adn, sizeof( adn ) ); 
     stntbi( &nd, lun, adn, nemo, cseq, sizeof( adn ), 8, 55 );
 
@@ -176,14 +176,14 @@ void stseq( f77int *lun, f77int *irepct, f77int *idn, char nemo[8],
 **		  Is nemo2 already listed as an entry within the internal
 **		  Table B?
 */
-		  nemtab( lun, nemo2, &pkint, tab, &iret, 8, 1 );
+		  nemtab( lun, nemo2, &pkint, tab, &iret, 8, (size_t)1 );
 		  if ( ( iret == 0 ) || ( tab[0] != 'B' ) ) {
 /*
 **		    No, so create and store a new Table B entry for nemo2.
 */
 		    tab[0] = 'B';
 		    tab[1] = '\0';
-		    nb = igetntbi( lun, tab, 1 );
+		    nb = igetntbi( lun, tab, (size_t)1 );
 
 		    if ( ix == 4 ) {
 			sprintf( rpseq, "Associated field of %3lu bits",
@@ -394,7 +394,7 @@ void stseq( f77int *lun, f77int *irepct, f77int *idn, char nemo[8],
 **	    Is cdesc[i] already listed as an entry in the internal Table B?
 */
 	    numtbd( lun, &cdesc[i], nemo2, tab, &iret, sizeof( nemo2 ),
-		    1 );
+		    (size_t)1 );
 	    if ( ( iret == 0 ) || ( tab[0] != 'B' ) ) {
 /*
 **		No, so search for it within the master table B.
@@ -403,7 +403,7 @@ void stseq( f77int *lun, f77int *irepct, f77int *idn, char nemo[8],
 /*
 ** 		Start a new Table B entry for cdesc[i].
 */
-		nb = igetntbi( lun, tab, 1 );
+		nb = igetntbi( lun, tab, (size_t)1 );
 		cadn30( &cdesc[i], adn2, sizeof( adn2 ) ); 
 		stntbi( &nb, lun, adn2, &MSTABS_BASE(cbmnem)[ipt][0],
 			&MSTABS_BASE(cbelem)[ipt][0], sizeof( adn2 ), 8, 55 );
